+++ internal/impl/snowflake/output_snowflake_put.go
@@ -22,7 +22,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/gofrs/uuid"
+	"github.com/gofrs/uuid/v5"
 	"github.com/golang-jwt/jwt/v5"
 	"github.com/snowflakedb/gosnowflake"